Fix keyboard shortcut conflict on Mac
Qt maps all instances of the "control" key in the code to the "command" key on
the Mac.  This is usually fine, but when the keyboard shortcut for pause
"control+space" turns into "command+space", it conflicts with the system-wide
keyboard shortcut for Apple's Spotlight utility.  This changes the code to use
the real "control+space" (what Qt calls "meta+space") for pause on the Mac.