A keyboard-based mechanism for moving the selection to nearby events, for use with tweak editing.
Note that determining the nearest event in a given direction is currently based
on Euclidean distance between the starts of the events (with MIDI time and line
number translated to display coordinates).  Testing shows that this does not
produce intuitive results in all cases, so the heuristic will need refinement.