fix excessive CPU use in the MIDI SenderThread
I ran a profiler to determine where the runaway CPU usage appeared, and it
showed up as dominated by calls to msleep() in SenderThread.  I did the
standard replacement of the busy loop with a mutex and wait condition, which
solved the problem.  Initial interactive testing showed no degradation of
timing accuracy in playback.