add subtle acceleration
Again borrowing from game design.  When doing a small increase or decrease on
properties with a large range (time, pitch bend, and tempo), make the increment
more than one if the user repeats the action immediately (via automatic key
repeat).  The amount of increment is proportional to how long the user holds
down the key, so it accelerates as you hold the key longer.