Add a medium decrease/increase level.
I found that two levels weren't sufficient for the properties with
larger value ranges (time, pitch bend, and tempo).  It's important to
keep the small increment as 1 so that the user can get precise values,
but that meant that the "large" would either be too large or too small
for some cases no matter where I set it.  Three levels allowed me to
spread them out better.