load metronome click audio file from resource
This means that the program doesn't need to look for the file in a particular
local path since it's embedded in the executable (it was already there anyway).
QMediaPlayer doesn't reliably support loading from resources but QSoundEffect
does, and the Qt docs recommend using the latter for this kind of purpose
anyway.