Clean up the order of tools in the menu and toolbar
The order of "current tool" choices is different in the toolbar, menu, and
keyboard bindings.  The toolbar also sticks things that aren't current tool
choices (align, equalize, and quantize -- you don't click on the MatrixWidget
to apply them) in the middle of the list of current tool choices.

This modification makes it easier for the user to understand by changing the
order to be consistent.  It puts the more basic things at the beginning of the
overall list and at the beginning of each section:  standard tool, pencil,
eraser, <separator>, select single, select box, select left, select right,
<separator>, move, move horizontally, move vertically, duration, <separator>,
measures, time signature, tempo, <separator>, align left, align right,
equalize, quantize, magnet.

It also adds a keybinding (escape, like in Cubase) and toolbar button for MIDI
panic.