"interpolate" now extrapolates too, which seems to fix it for real
The two pass approach didn't help, and made lines which were drawn straight
actually appear curved when you released the mouse.  The problem actually
appeared to be coming from needing to handle points a little beyond the end of
the "toAlign" array, so I had to make the interpolate function extrapolate off
the right end of that line.  Also fixed a problem with inserting lots of
"minimum possible value" events if the user released the mouse below the bottom
edge of the window.