now properly working for adding events, but still not removing old ones correctly
In particular, I still need to figure out what the righthand bound of the range
should be for removing the old events.  maxTick isn't correct, even if you
recompute it as xPosToTock(toAlign.last().first) each time you go around the
loop, because the end point depends on the tempo events you're going to insert
not the ones which are already there.  I may need to intersperse the remove and
add rather than having them in separate loops.