now the time range of adds and deletes is accurate in all cases
What you draw is evaluated in the context of what's already there, so that
means that a tempo line which you draw straight will end up curved once you
release the mouse.  That's unintuitive, but I'm starting to think it's not
actually a bug, just an accurate reflection of what the straight line really
means.