pasteAction(): Iterate events backwards
When pasting events, they may extend beyond the current duration of the
file. In this case the file duration is increased. When iterating from
the first to last event, within each iteration the file duration is
incremented only by a fraction of the final duration. Thus, it is
generally much faster to start with the last event and iterate
backwards.