Fix unnecessary copying of midi channels
When inserting midi events, the channel corresponding to each event is
copied into the protocol stack. Currently, the channel is copied for
each single inserted event. However, this is unnecessary and this leads
to a massive increase in allocated memory if many events are to be
inserted. It is only necessary to copy all channels involved once, but
not for each event separately.